rl: deduplicate  code
